+++ packages/analytics-angular/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.3](https://github.com/Availity/sdk-js/compare/@availity/analytics-angular@3.0.2...@availity/analytics-angular@3.0.3) (2019-04-17)
+
+**Note:** Version bump only for package @availity/analytics-angular
+
+
+
+
+
+## [3.0.2](https://github.com/Availity/sdk-js/compare/@availity/analytics-angular@3.0.1...@availity/analytics-angular@3.0.2) (2019-04-17)
+
+**Note:** Version bump only for package @availity/analytics-angular
+
+
+
+
+
+## [3.0.1](https://github.com/Availity/sdk-js/compare/@availity/analytics-angular@3.0.0...@availity/analytics-angular@3.0.1) (2019-04-17)
+
+**Note:** Version bump only for package @availity/analytics-angular
+
+
+
+
+
 # [3.0.0](https://github.com/Availity/sdk-js/compare/@availity/analytics-angular@2.6.4...@availity/analytics-angular@3.0.0) (2019-03-18)